feat: freeze the scale between years

+++ app/src/components/Charts/StreamPerMonth/query.tsx
@@ -14,7 +14,7 @@ export function queryByYear(year: number | undefined) {
       SELECT
         LAST_DAY(ts::DATE) as ts,
         SUM(ms_played) as ms_played,
-        COUNT(*)::INTEGER as count_streams
+        COUNT(*) as count_streams
       FROM ${TABLE}
       ${year ? `WHERE YEAR(ts::DATE) = ${year}` : ''}
       GROUP BY LAST_DAY(ts::DATE)
@@ -22,7 +22,7 @@ export function queryByYear(year: number | undefined) {
     SELECT
       MAKE_DATE(YEAR(all_months.ts), MONTH(all_months.ts), 1) as ts,
       COALESCE(monthly_streams.ms_played, 0) as ms_played,
-      COALESCE(monthly_streams.count_streams, 0) as count_streams
+      COALESCE(monthly_streams.count_streams, 0)::INT as count_streams
     FROM all_months
     LEFT JOIN monthly_streams ON all_months.ts = monthly_streams.ts
     ORDER BY all_months.ts`
